Update with to be on part with version used internally. This splits out functionality into separate modules.
Among major user-visible changes is addition of simple media-server
functionality, which allows on-demand replacing one of RTP streams in session
with one generated from local file. This will be used to introduce
music-on-hold and similar functionality in nathelper later.

Protocol remains backward-compatible, so that version reported by rtpproxy
so far remains the same.

Changes have been tested on FreeBSD and Linux, users of other arches are
advised to test and report if there are any problems.