Merge in latest changes from the BSD version:
o Implement ability to configure port range via command line using
two new options: -m port_min and -M port_max. Inspired by the OpenSER
ticket #1707122.

o Complete moving global variables to the struct cfg. This should not
have any vissible effect, but allows us to move functions our of main.c
easily.

o Remove usage of signals for running infrequent periodic tasks (such
as time-to-live tracking). Instead, run it explicitly at the end of the
main loop. This should provide better visibility of what's going on
as well as allow us to move to the multi-thread design in the future
more easily. Not even to mention that we are shaving two system calls
per loop.

o Fix rtp_samples2bytes() and rtp_bytes2samples() for the G.723.1 codec.
For now only support 6.3kbps version, since the interface within rtpproxy
itself is not flexible enough to support both, this means that for now
only this version of G.723.1 will be resized. We will improve the
algorythm so that variable-bitrate codecs could be supported as well.