Improve handling of session updates (i.e. re-INVITEs). When the update comes in check if historically address supplied by the end point can be trusted and update if so, otherwise just set flag that the next RTP packet can update recorded RTP address, but leave the address intact.
This should fix issues with T.38, when re-INVITE comes from NATed client,
we update recorded address with its local IP and start sending packets
to that incorrect address. Since T.38 is not symmetric, we may end up
receiving no packets from the NATed peer in this situation (it waits
(it waits to the other party to start sending indefinitely).