Clean up two last changes from dpocock:
- Fix usage of rtpp_log_write;

- alocate timeout connecting socket during initial initialization sequence.
It can be reused by simply invoking connect() over and over again. This
should provide better protection against run-time resource exhaustion.

- use asserts in order to detect logic errors in program flow, don't
tolerate them by having checks that are always false.

- properly handle BSD-like system, where struct sockaddr_un has sun_len
member that has to be inited appropriately.

- AF_UNIX -> AF_LOCAL. They are the same, but later is more posixly correct
as well as it's consistent to what we use in other parts of the code.

- interpret argument to the -n command not starting with unix: as pathname
instead of rejecting it. This behaviour is in line with the -s option.

- Handle EINTR condition when doing send(2).

- move two functions out from main.c to rtpp_session.c where they belong.
Some time ago we did lot of work to trim down bloated main.c, let's not
unto that.

- use warn/err/errx instead of fprintf/perror.

- put space after `if' for consistency.

- move notify_buf inside struct rtpp_timeout_handler.

- capitalize enum values, so that it's clear they are not local variables.