o Extent 'U' / 'L' commands with optional payload types list. When specified, this list is saved in the session properties and can be used later on in the 'P' command;
o extend 'P' command so that when codec list argument is the word
"session" (without quotes) then the list of payload types saved by
the 'U' or 'L' will be used to select the most appropriate prompt.

o Add a new protocol version 20081102, so that we can use this
functionality from nathelper when available.

This is part of the work to enable playing announcements and MOH
from the

Sponsored by:	Telio Telecom AS