Fix serious issue wich can happen with default loging method (syslog) at the default log level (debug). I'm unsure if it affects all Linux systems or not, but at least on one CentOS release 5.3 the vsyslog(3) call takes ridiculously large amount of time to complete. Few of such calls easily take 1-2 seconds. Since the RTPproxy is still single-threaded, this means that traffic halts for that duration, which is obviously bad.
Run vsyslog(3) calls in the separate working thread instead. This
ensures that logging won't interfere with normal activity even if the
I/O system is loaded or when logging is done over network.

While I am here, set default logging level when running in background
to "WARN", since the current level ("DBUG") is too verbose to be a
default.