Re-do locking around command-line socket so that it actually works. Instead of locking cf structure back and forth, lock it for the duration of the handle_command() call and only unlock it when it's safe for the duration of the potentially blocking call, like resolve().
Also split handle_command() into two sections - one that is lock-free
that reads on the command socket, and the second that actually does
processing. While not ideal, this should prevent command handling to
produce any stalls in the actual packet forwarding.

The next easy step would be to allow several commands to proceed
independently. Also, provide separate locking for the cf elements
that are used by the poll(2), so it's possible to process commands
while polling.