Make stdio: command mode actually useful. Instead of dumping the output into the same file descriptor as input, which only works for terminal i/o, actually dump output to the stdout, while still taking input from the stdin.
Also improve stream-based communication support to accept more than
one command in the batch, don't expect sender to pause and wait for
the reply after issuing a command.

And last but not least, gracefully close the rtpproxy if stdio is
the only command input when it's closed, so that we can run something
like:

rtpproxy -s stdio: -f < some.commands

this is going to be quite useful for automated regression testing.