Make session latching more intelligent. Instead of disallowing another latch within certain interval after previous one (UPDATE_WINDOW), allow another update any time if new packet meets criterias:
1. It's valid RTP packet (as far as rtp_packet_parse() can tell);

2. It has the same SSRC value that has been saved during the
   previous latching;

3. Its sequence number is greated than the sequence number that
   we saved while doing the previous latching.

If any of (1)-(3) fails, then we revert to the old way with
disallowing new update within UPDATE_WINDOW after the previous
one.

This fixes the issue when endpoint generates 3 packets from
the proper IP, but then switches to completely different port
and sticks with that for the rest of the session. Right now,
the rtpproxy would latch to the first packet out of those 3 and
discard the rest considering those a garbage. We are not 100%
sure if it's endpoint's fault or just some NAT stupidity.

Swill somewhat WiPish, as we need to create a proper test for this
scenario.

Also, add printing of SSRC and initial sequence number while
doing latching. It is helpful to identify session in the log and
could also be useful in some other cases.