Implement wildcard notification sockets, which basically allows us to deliver notifications to any number of client nodes over network without the need to enumerate each of those nodes separately.
How it works is basically when notification socket is specified as
tcp:%%CC_SELF%%:2121, in both argument to the rtpproxy itself and
in the appropriate U / L command, the proxy would take source address
that the command has been received from (say IP 1.2.3.4) and
substitute %%CC_SELF%% with that IP, delivering notification
to the host 1.2.3.4, port 2121.

As it being said, it allows any number of nodes to be notified
in easily configurable and securable manner.