Further refine and refcount timed API and also make it easier to use. Split it into two main cases - simple tasks that don't return anything and are destroyed on exit and more advanced ones that actually return handler.
Simplify ability to run periodic tasks by allowing return value from
the timer handler with either CB_MORE or CB_LAST. In the former case
the timer will be re-queued to run at constant interval.

Allow anyone supplying us with the callback function also provide a refcount
to the underlying argument pointer to that function, so that we keep it
alive while the callback is pending.

Simplify cancelling by storing reference to the parent rtpp_timed in the
"task" sub-object.