o Relax rematching algorythm a bit to disregard seq numbers mismatches if we are within update window.
o add what seems to be a missing call to _rtpp_stream_latch() in one of the
code branches. I thing I've seen some compiler warnings on some of the arches.

For now those are somewhat WiP, required to run recording changes. I'll
do a cherry-pick into experimental branch and see how it fares in our battery
of tests.