Add DTLS encrypt/decrypt module, with crypto parts based loosely off the basesip's and SRTP decryption done by the libsrtp2.
Add machinery to setup/destroy SSL/DTLS context.

Add code to setup DTLS connection.

Implement network I/O to and from TLS code.

Add support for active mode and simple FSM to track state of
the connection.

Implement basic control protocol allowing to supply mode
of remote "A" - for active, "P" for passive. For the case
when we are initiating session the "S" (setup) is
provided, roughly equivalent of "actpass". The "A" and "P"
take the SHA-256 cert fingerprint and SSRC when available
and return either success or failure, "S" takes no arguments
and sends back our SHA-256 digest in response.

Two more commands are provided to detach/destroy DTLS context:
"D" and "U". "D" is to be used if we have received DTLS session,
"U" if we have initiated.

Only build dtls_gw when BUILD_CRYPTO is set.