Add flag for RNG re-seed to help with brute-forcing cryptos in fuzz_rtp_session.
Call __afl_manual_init() early in init if available.

Parse fuzzer input as a simple variable-length encoded
data, so that one input might result in multiple commands
executed sequentially. This is expected to let fuzzer to
hit more interesting scenarious easier since it's not
inter-inputs stateful. As well as to hopefully make
reproduction simpler as well.