Rework refcounter logic to streamline deallocations.
Instead of having destructor and "pre-destructor" and a
complicated logic around which is set and not, have an
array of deallocator / poiner pairs of a fixed max size
and allow application to push onto this stack.

On deallocation time execute the functions in LIFO-manner.

This allows to untangle and de-complicate deallocation
logic in many scenarios.