Expose GET_SMETHOD() to retrieve a method reference.
Add RTPP_OBJ_BORROW() macro, useful when we need to grab a
reference to an object A and hold it for a lifecycle of the
object B, so this would increase reference and add attach
a destructor to deref when B is deallocated. Care should
be taken to observe relative order when multiple objects
are borrowed.